app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.1
openai==1.51.2
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2

pandas==2.2.3
openpyxl==3.1.5
jinja2==3.1.4
python-multipart==0.0.12
rapidfuzz==3.9.6

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
